+++ packages/localize-tools/testdata/extract-xlb/goldens/foo.ts
@@ -7,9 +7,9 @@
 import {html} from 'lit-html';
 import {msg} from '@lit/localize';
 
-// msgdesg: Description of Hello World
+/** @desc Description of Hello World */
 msg('Hello World!');
 
 const name = 'friend';
-// msgdesc: Description of Hello $(name)
+/** Description of Hello $(name) */
 msg(html`Hello <b>${name}!</b>`);